Complete update of RAG metrics
